update manage product and shop by nsx

+++ FurniMatch.Api/DTOs/ProductDto.cs
@@ -18,8 +18,8 @@ namespace FurniMatch.Api.DTOs
 
         public bool CustomSizeSupported { get; set; }
 
-        public int? Length { get; set; }
-        public int? Width { get; set; }
-        public int? Height { get; set; }
+        public double? Length { get; set; }
+        public double? Width { get; set; }
+        public double? Height { get; set; }
     }
 }
