+++ FurniMatch.Api/Models/QuotationRequest.cs
@@ -20,9 +20,9 @@ namespace FurniMatch.Api.Models
         [MaxLength(100)]
         public string? Pattern { get; set; }
 
-        public int Length { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double Length { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
 
         [MaxLength(100)]
         public string? Material { get; set; }
